+++ src/main.ts
@@ -12,6 +12,7 @@ import { renderCampaignMap } from './ui/CampaignMapScreen';
 import { renderCombat } from './ui/CombatScreen';
 import { renderBattleResults } from './ui/BattleResultsScreen';
 import { renderCharacterSheet } from './ui/CharacterSheetScreen';
+import { renderInventory } from './ui/InventoryScreen';
 
 // Development: Make test functions available in console
 import {
@@ -69,9 +70,9 @@ if (!app) {
   ScreenManager.registerScreen('combat', renderCombat);
   ScreenManager.registerScreen('battleResults', renderBattleResults);
   ScreenManager.registerScreen('characterSheet', renderCharacterSheet);
+  ScreenManager.registerScreen('inventory', renderInventory);
   
-  // TODO: Register other screens as we build them
-  // ScreenManager.registerScreen('inventory', renderInventory);
+  // TODO: Register Settings screen when built
   // ScreenManager.registerScreen('settings', renderSettings);
   
   // Event listeners